moved cmd to a separate package (#29)
