+++ main.go
@@ -1,0 +1,8 @@
+package main
+
+import "github.com/defenseunicorns/uds-security-hub/cmd"
+
+// main function remains to call Execute.
+func main() {
+	cmd.Execute()
+}
